+++ src/textual_fspicker/file_open.py
@@ -33,6 +33,7 @@ class FileOpen(BaseFileDialog):
         must_exist: bool = True,
         default_file: str | Path | None = None,
         double_click_directories: bool = True,
+        suggest_completions: bool = True,
     ) -> None:
         """Initialise the `FileOpen` dialog.
 
@@ -45,6 +46,7 @@ class FileOpen(BaseFileDialog):
             must_exist: Flag to say if the file must exist.
             default_file: The default filename to place in the input.
             double_click_directories: Double click to open directories.
+            suggest_completions: Should the `Input` suggest completions?
 
         Notes:
             `open_button` and `cancel_button` can either be strings that
@@ -59,6 +61,7 @@ class FileOpen(BaseFileDialog):
             filters=filters,
             default_file=default_file,
             double_click_directories=double_click_directories,
+            suggest_completions=suggest_completions,
         )
         self._must_exist = must_exist
         """Must the file exist?"""