+++ ChangeLog.md
@@ -11,6 +11,9 @@
 - Added `textual_fspicker.Icons` so that the user of the library can set
   their own icons that are displayed next to directory entries.
   ([#74](https://github.com/davep/textual-fspicker/pull/74))
+- Added support for suggested completions when typing paths into the
+  `FileOpen` and `FileSave` dialogs.
+  ([#76](https://github.com/davep/textual-fspicker/pull/76))
 
 ## v0.5.0